requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
ale-py=0.8.0=pypi_0
argparse=1.4.0=pypi_0
attrs=22.2.0=pypi_0
autorom=0.4.2=pypi_0
autorom-accept-rom-license=0.5.4=pypi_0
bitmath=1.3.3.1=pypi_0
blas=1.0=mkl
box2d-py=2.3.5=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2023.5.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
contourpy=1.0.7=pypi_0
cryptography=38.0.4=py38h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.140=0
cuda-documentation=12.0.140=0
cuda-driver-dev=11.7.99=0
cuda-gdb=12.0.140=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.140=0
cuda-nsight-compute=12.0.1=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.140=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.146=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.146=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.140=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pypi_0
cython=0.29.33=pypi_0
decorator=4.4.2=pypi_0
dill=0.3.6=pypi_0
dm-env=1.6=pypi_0
dm-tree=0.1.8=pypi_0
enum-tools=0.9.0.post1=pypi_0
envpool=0.8.1=pypi_0
evostra=2.5.2=pypi_0
fasteners=0.18=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py38h06a4308_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
gds-tools=1.5.1.14=0
giflib=5.2.1=h5eee18b_1
glfw=2.5.6=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
gymnasium=0.27.1=pypi_0
gymnasium-notices=0.0.1=pypi_0
hbutils=0.8.1=pypi_0
idna=3.4=py38h06a4308_0
imageio=2.25.0=pypi_0
imageio-ffmpeg=0.4.8=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.10.2=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jax-jumpy=0.2.0=pypi_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufft-dev=10.7.2.124=h98a8f43_0
libcufile=1.5.1.14=0
libcufile-dev=1.5.1.14=0
libcurand=10.3.1.124=0
libcurand-dev=10.3.1.124=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=hecacb30_0
libtorrent=2.0.7=pypi_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4=4.3.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.6.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
moviepy=1.0.3=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py38h06a4308_0
mujoco=2.2.1=pypi_0
mujoco-py=2.1.2.14=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py38h06a4308_1
nsight-compute=2022.4.1.6=0
numpy=1.24.1=pypi_0
numpy-base=1.23.5=py38h31eccc5_0
opencv-python=4.7.0.68=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pillow=9.3.0=py38hace64e9_1
pip=22.3.1=py38h06a4308_0
pluggy=1.0.0=pypi_0
proglog=0.1.10=pypi_0
py=1.11.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygame=2.1.0=pypi_0
pygments=2.14.0=pypi_0
pyopengl=3.1.6=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
pytest=7.0.1=pypi_0
python=3.8.16=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python-graphviz=0.20.1=pypi_0
pytimeparse=1.1.8=pypi_0
pytorch=2.0.1=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=py38h06a4308_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h5082296_0
swig=4.1.1=pypi_0
sympy=1.11.1=py38h06a4308_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torchaudio=2.0.2=py38_cu117
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu117
tqdm=4.64.1=pypi_0
treevalue=1.4.4=pypi_0
types-protobuf=4.21.0.4=pypi_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.14=py38h06a4308_0
uuid=1.30=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
zipp=3.12.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
